// src/lib/api/api.ts
export interface components {
    schemas: {
// ...
        DStaff: {
            /** Format: int64 */
            id?: number;
            isActive?: boolean;
            /** Format: date-time */
            createdDate?: string;
            /** Format: date-time */
            updatedDate?: string;
            phoneNumber?: string;
            /** Format: float */
            salary?: number;
            /** @enum {string} */
            currency?: "PEN" | "USD";
            /** Format: date-time */
            hireDate?: string;
            /** @enum {string} */
            role?: "SALES" | "COUNTER" | "ACCOUNTING" | "OPERATIONS" | "SUPERADMIN" | "SUPPORT";
            /** Format: int64 */
            userId?: number;
            user?: components["schemas"]["DUser"];
        };
// ...
        DCustomer: {
            /** Format: int64 */
            id?: number;
            isActive?: boolean;
            /** Format: date-time */
            createdDate?: string;
            /** Format: date-time */
            updatedDate?: string;
            firstName?: string;
            lastName?: string;
            email?: string;
            phoneNumber?: string;
            /** Format: date */
            birthDate?: string;
            /** @enum {string} */
            idDocumentType?: "PASSPORT" | "DNI" | "DRIVER_LICENSE" | "RUC" | "CE";
            idDocumentNumber?: string;
            address?: string;
            nationality?: string;
        };
// ...
        DIncidency: {
            /** Format: int64 */
            id?: number;
            isActive?: boolean;
            /** Format: date-time */
            createdDate?: string;
            /** Format: date-time */
            updatedDate?: string;
            reason?: string;
            /** Format: float */
            amount?: number;
            /** Format: date-time */
            incidencyDate?: string;
            /** @enum {string} */
            incidencyStatus?: "PENDING" | "APPROVED" | "REJECTED";
            /** Format: int64 */
            liquidationId?: number;
            pending?: boolean;
            approved?: boolean;
        };
        DLiquidation: {
            /** Format: int64 */
            id?: number;
            isActive?: boolean;
            /** Format: date-time */
            createdDate?: string;
            /** Format: date-time */
            updatedDate?: string;
            /** Format: float */
            currencyRate?: number;
            /** Format: float */
            totalAmount?: number;
            /** Format: date-time */
            paymentDeadline?: string;
            /** Format: int32 */
            companion?: number;
            /** @enum {string} */
            status?: "IN_QUOTE" | "PENDING" | "ON_COURSE" | "COMPLETED";
            payments?: components["schemas"]["DPayment"][];
            /** @enum {string} */
            paymentStatus?: "PENDING" | "ON_COURSE" | "COMPLETED";
            flightServices?: components["schemas"]["DFlightService"][];
            hotelServices?: components["schemas"]["DHotelService"][];
            tourServices?: components["schemas"]["DTourService"][];
            additionalServices?: components["schemas"]["DAdditionalServices"][];
            /** Format: int64 */
            customerId?: number;
            customer?: components["schemas"]["DCustomer"];
            /** Format: int64 */
            staffId?: number;
            staffOnCharge?: components["schemas"]["DStaff"];
            incidencies?: components["schemas"]["DIncidency"][];
            overdue?: boolean;
            /** Format: float */
            remainingAmount?: number;
            /** Format: float */
            totalPaid?: number;
        };
        DPayment: {
            /** Format: int64 */
            id?: number;
            isActive?: boolean;
            /** Format: date-time */
            createdDate?: string;
            /** Format: date-time */
            updatedDate?: string;
            /** @enum {string} */
            method?: "DEBIT" | "CREDIT" | "YAPE" | "OTHER";
            /** Format: float */
            amount?: number;
            /** Format: int64 */
            liquidationId?: number;
            /** @enum {string} */
            validationStatus?: "PENDING" | "VALID" | "INVALID";
            valid?: boolean;
            pending?: boolean;
        };
        DTour: {
            /** Format: int64 */
            id?: number;
            isActive?: boolean;
            /** Format: date-time */
            createdDate?: string;
            /** Format: date-time */
            updatedDate?: string;
            /** Format: date-time */
            startDate?: string;
            /** Format: date-time */
            endDate?: string;
            title?: string;
            /** Format: float */
            price?: number;
            place?: string;
            /** @enum {string} */
            currency?: "PEN" | "USD";
            /** @enum {string} */
            status?: "PENDING" | "COMPLETED" | "CANCELED";
            /** Format: int64 */
            tourServiceId?: number;
            pending?: boolean;
        };
// ...
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}